% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/but-package.R
\docType{package}
\name{but-package}
\alias{but-package}
\title{but: Edit and Expand Existing Functions}
\description{
Create functions based on existing functions, but with modified formals or additional processing of their inputs or output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/but}
  \item Report bugs at \url{https://github.com/hrryt/but/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
