% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{do}
\alias{do}
\title{Execute a Function Call}
\usage{
do(.f, ...)
}
\arguments{
\item{.f}{a function}

\item{...}{passed to \code{.f}}
}
\value{
The result of \code{.f(...)}.
}
\description{
Constructs and executes a function call from a function and extra arguments.
}
\examples{
matrix |> but(ncol = nrow) |> do(0, nrow = 2)
sapply(c(mean, median, IQR), do, print(runif(5)))

}
\seealso{
\code{\link[=but]{but()}}, \code{\link{|>}}.
}
