% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/but.R
\name{but}
\alias{but}
\title{Modify a Function's Formals, Inputs and Output}
\usage{
but(
  .f,
  ...,
  .first = FALSE,
  .nse = FALSE,
  .wrap = TRUE,
  .store = FALSE,
  .pass_all = TRUE
)
}
\arguments{
\item{.f}{a function (a primitive or a closure, i.e., 'non-primitive')}

\item{...}{modified formals and instructions for pre- and post-processing}

\item{.first}{should supplied formals come first, and in the order specified?}

\item{.nse}{should the \strong{n}on-\strong{s}tandard \strong{e}valuation of \code{.f} be accounted for
using \code{\link[=match.call]{match.call()}}?}

\item{.wrap}{should \code{.f} be wrapped in a new function or its formals modified directly?}

\item{.store}{should \code{.f} be stored in the environment of the returned function?}

\item{.pass_all}{should arguments not present in the formals of \code{.f} be passed to \code{.f}
if it has \link{dots} to absorb them?}
}
\value{
A function.
}
\description{
Constructs a function that calls the input function with modified formal arguments
and optionally additional processing of its inputs and output.
}
\details{
This documentation assumes \code{.f} is a function, but it can be a call, in which
case \code{but} is applied to the function of the call before evaluation of the call.
\code{but(f(x), y = x)} is equivalent to \code{but(f, y = x)(x)}.

If \code{.f} is a primitive without a well-defined argument list,
its formals are assumed to be \code{alist(... = )} and a warning is given.

If \code{.nse} is \code{TRUE}, a call to \code{.f} is constructed from \code{match.call()}
in the body of the returned function. This is useful if \code{.f} is, for example,
a modelling function that captures its call and displays it when printing.
}
\section{Named arguments}{


Each named argument supplied to \code{but()} replaces the default value of the
argument of \code{.f} with that name, or if it is not present in the formals of \code{.f},
is added to the end of the argument list.

A formal argument can be removed by quoting \code{.rm}, e.g. \code{but(.f, x = .rm)}.
}

\section{Unnamed arguments}{


Unnamed arguments supplied to \code{but()} are used as language objects
to build the body of the returned function.

If an unnamed argument references \code{.out}, e.g. \code{but(.f, g(.out))},
it is appended to the body after the call to \code{.f} is assigned to \code{.out},
such that the output of \code{.f} can be modified before it is returned.

Otherwise, unnamed arguments are added to the body before the call to \code{.f},
for example so that arguments can be modified before being passed to \code{.f}.

The walrus operator \verb{:=} can be used to modify the call to \code{.f} directly,
regardless the value of \code{.nse}. For example, \code{but(.f, x := y)} might be
equivalent to \code{function(...) .f(..., x = y)}.
}

\examples{
# named arguments specify default values to change
(max_rm <- max |> but(na.rm = TRUE))
max_rm(0, NA, 2, 1)
(x <- log(c(0, NA, 1)))
min(x)
# unnamed arguments are added to the body of the returned function
(min_inf <- min |> but(if(-Inf \%in\% c(...)) return(-Inf)))
min_inf(x)
# applying but directly to a call saves dealing with the intermediate function
min(x) |> but(if(-Inf \%in\% c(...)) return(-Inf))

args(lm)
# leaving data's value missing ensures there is no default value
# .first = TRUE puts data to the front of the argument list
(lm4pipe <- lm |> but(data = , .first = TRUE))
# non-standard evaluation in the body of lm makes wrapping difficult
try(mtcars |> subset(cyl == 4) |> lm4pipe(mpg ~ disp))
# use .wrap = FALSE to modify the formals of lm directly instead
lm4pipe <- lm |> but(data = , .first = TRUE, .wrap = FALSE)
mtcars |> subset(cyl == 4) |> lm4pipe(mpg ~ disp)

resample <- function(x) x[sample(nrow(x), replace = TRUE), , drop = FALSE]
# wrapping that respects lm's NSE can be achieved with .nse = TRUE
# this allows the use of := to specify what the call to lm should look like
(lm4pipe_resample <- lm |> but(
  data = , .first = TRUE, .nse = TRUE,
  data := resampled_data, resampled_data <- resample(data)
))
mtcars |> subset(cyl == 4) |> lm4pipe_resample(mpg ~ disp)

(m <- diag(4))
m[2, ] # `[` defaults to drop = TRUE
# .store = TRUE ensures the initial `[` function is stored as .f
(`[` <- `[` |> but(drop = FALSE, .store = TRUE))
environment(`[`)$.f
m[2, ] # `[` now defaults to drop = FALSE
m[2, , drop = TRUE]
rm(`[`)

(none <- Negate(any))
# unnamed arguments that reference .out are added after the call to .f
(none <- any |> but(!.out))
strsplit |> but(.out[[1]])
subset |> but(drop = TRUE, droplevels(.out))
# use on.exit() if post-processing does not reference .out
read.csv |> but(on.exit(unlink(file)))

# use .rm to remove arguments
(square <- matrix |> but(ncol = .rm, ncol <- nrow, nrow = sqrt(length(data))))
# use := to modify the call directly
(square <- matrix |> but(ncol = .rm, ncol := nrow, nrow = sqrt(length(data))))
square(1:9, byrow = TRUE)
square(1, 3)

# use := .rm to remove arguments from the call
read.csv |> but(
  keep.white = TRUE, strip.white := !keep.white, keep.white := .rm
)
# use .pass_all = FALSE to avoid passing
# extra arguments to the dots of .f
read.csv |> but(
  keep.white = TRUE, strip.white := !keep.white, .pass_all = FALSE
)

(cb1 <- cbind |> but(`:=`(, 1))) # := missing first argument
cb1(1:6, 1:2)

}
\seealso{
\code{\link{|>}}.
}
